remerge for IWCD new geom branch to compile correctly

+++ include/WCSimWCAddDarkNoise.hh
@@ -18,6 +18,7 @@ class WCSimWCAddDarkNoise : public G4VDigitizerModule
 public:
   
   WCSimWCAddDarkNoise(G4String name, WCSimDetectorConstruction*, G4String);
+  //WCSimWCAddDarkNoise(G4String name, WCSimDetectorConstruction*, G4String);
   ~WCSimWCAddDarkNoise();
   
 public:
@@ -31,9 +32,9 @@ public:
   void SetConversion(double iconvrate){ ConvRate = iconvrate; }
   void SetDarkMode(int imode){DarkMode = imode;}
   int GetDarkMode(){return DarkMode;}
-  void SetDarkHigh(double idarkhigh){DarkHigh = idarkhigh;}
+  void SetDarkHigh(int idarkhigh){DarkHigh = idarkhigh;}
   double GetDarkHigh(){return DarkHigh;}
-  void SetDarkLow(double idarklow){DarkLow = idarklow;}
+  void SetDarkLow(int idarklow){DarkLow = idarklow;}
   double GetDarkLow(){return DarkLow;}
   void SetDarkWindow(int idarkwindow){DarkWindow = idarkwindow;}
   int GetDarkWindow(){return (int)(DarkWindow);}
@@ -54,7 +55,7 @@ private:
 
   WCSimDetectorConstruction* myDetector;
   G4String detectorElement;
-
+  
   std::vector<std::pair<double, double> > ranges;
   std::vector<std::pair<double, double> > result;
